Refactor LED handling and configuration
- Changed the data type of `num_leds` from `uint16_t` to `uint32_t` in multiple files to accommodate larger LED counts.
- Updated `MAX_LEDS_PER_REGION` from 7 to 15 in `JaamConfig.h` to allow for more LEDs per region.
- Introduced legacy handling for LED brightness mapping in `JaamLed.cpp` based on the `legacy` variable.
- Added a new function `getRegionEntryLegacy` to manage region entries based on the legacy mode.
- Modified various functions to use the new `uint32_t` type for LED counts and indices.
- Enhanced the web interface to include a dropdown for selecting the firmware mode (`legacy`).
- Updated logging statements to be conditionally compiled for debugging purposes.

+++ src/JaamAnimation.cpp
@@ -753,7 +753,7 @@ void AnimationManager::cleanupAnimation(AnimationParams* anim, int index) {
     LOG.printf("[ANIMATION] Cleaned up animation slot %d, active count: %d\n", index, activeCount);
 }
 
-std::vector<FreeLedInfo> AnimationManager::getFreeLeds(Adafruit_NeoPixel* strip, uint16_t num_leds) {
+std::vector<FreeLedInfo> AnimationManager::getFreeLeds(Adafruit_NeoPixel* strip, uint32_t num_leds) {
     std::vector<FreeLedInfo> freeLedsResult;
     std::set<int> animatedLeds;
 
@@ -769,7 +769,7 @@ std::vector<FreeLedInfo> AnimationManager::getFreeLeds(Adafruit_NeoPixel* strip,
         xSemaphoreGive(animMutex);
     }
 
-    for (uint16_t j = 0; j < num_leds; ++j) {
+    for (uint32_t j = 0; j < num_leds; ++j) {
         if (animatedLeds.find(j) == animatedLeds.end()) {
             freeLedsResult.push_back({(int)j});
         }
